feat(#179): float32 triangle gauges, blockGaugeDeltas application, HUD floor
- PlayerState fire/water/woodGauge uint8 → float32 so tier-reduced block
  deltas (1/2^tier) accumulate fractionally; shadowGauge stays uint8.
- PlayerDuelist HUD floors the gauge for the numeric display; status `!`
  styling still triggers on the raw float >= STATUS_THRESHOLD.
- StatusEffects: verified PlayerLike gauges are `number` (float-safe);
  added fractional threshold tests (3.75 false, 4.0 true, 4.25 true).
- BattleRoom adjustGauge/clearAllGauges already float-correct (A3
  blockGaugeDeltas consumption + clampTo floor/soft-cap) — verified, no change.
- gauge-four-case E2E: Tier-0 delta=1.0 === 1, assertions still hold; noted.

+++ server/src/schemas/PlayerState.ts
@@ -19,10 +19,12 @@ export class PlayerState extends Schema {
 
   // Elemental gauges — one per triangle element (FIRE/WATER/WOOD). Fill when an
   // attack with that triangle component lands uncontested. Wind/Earth have no
-  // gauge (GDD §7.1).
-  @type('uint8') fireGauge: number = 0;
-  @type('uint8') waterGauge: number = 0;
-  @type('uint8') woodGauge: number = 0;
+  // gauge (GDD §7.1). float32 (not uint8) so tier-reduced block deltas
+  // (delta = 1/2^tier, e.g. 0.25 at Tier 2) accumulate fractionally; the HUD
+  // floors for display and status thresholds compare the raw float (#179, C6).
+  @type('float32') fireGauge: number = 0;
+  @type('float32') waterGauge: number = 0;
+  @type('float32') woodGauge: number = 0;
   // Shadow gauge (#134, GDD §7.1/§3.5). Fills like the triangle gauges (four-case
   // model) but caps at 5 and triggers Blinded at any stack (≥ 1). Shadow sits
   // outside the triangle, so it is tracked separately here.
